Refactor(hal): 송신 경로 전환 완결 및 어댑터 제거 (B3c)
B1~B3b 로 옮겨 온 SPI 패킷 버퍼 8비트 전환을 마무리한다.

변경
- dfu/tdc_dfu_ble_boot.c, dfu/tdc_dfu_ble_ota.c, sys/tdc_sys_error.c
  마지막 남은 로컬 조립 버퍼 3개와 호출 3곳 전환
- hal/tdc_hal_spi.c 의 전환용 어댑터 제거
  int* 를 받아 원소별 변환 복사하던 구 API 를 삭제했다
- tdc_hal_spi_write_tx_buffer_u8 을 원래 이름 tdc_hal_spi_write_tx_buffer
  로 되돌렸다. 전 호출부 57곳 일괄 정리
- 최종 시그니처: void tdc_hal_spi_write_tx_buffer(const uint8_t*, int)

전환 완결 확인
- 잔여 int 조립 버퍼 0건
- 잔여 _u8 접미사 0건
- HAL 버퍼 3개 전부 uint8_t (SPI_Rx_Buffer / SPI_Tx_Buffer 는 aligned(4))
- get_rx_packet_addr 반환형 uint8_t*

누적 효과 (B1 이전 대비)
- 정적 메모리: SPI 버퍼 3개가 248 -> 62 바이트 (186 바이트 절감)
- 코드 크기: 185592 -> 184144
- 패킷을 int 배열로 다루던 구조가 바이트 배열로 정리되어, 프로토콜
  문서의 바이트 오프셋과 코드 인덱스가 1:1 로 대응하게 됐다

검증
- CLI 빌드 에러 0 / 경고 0
- 실기 확인 대기 (dfu/ota 경로 포함 전 시나리오)

+++ src/2__cm3/source/dfu/tdc_dfu_ble_boot.c
@@ -13,7 +13,7 @@
 
 static void _send_resp_packet_boot(uint8_t* packet_data, uint8_t packet_len)
 {
-    int spi_buffer[BLE_DataPacketSize];
+    uint8_t spi_buffer[BLE_DataPacketSize];
     int spi_len;
 
     spi_len = packet_len;
